requirements.txt
python-dateutil
beautifulsoup
